requirements.txt
django-admin-tools
django-tagging
django-tagging-autocomplete

